Move Matchers class to the SAGESharp.Testing namespace
