+++ SAGESharpTests/SAGESharpTests.csproj
@@ -77,7 +77,6 @@
     <Reference Include="System.Xml.Linq" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Matchers.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AbstractEqualityByRefTests.cs" />
     <Compile Include="AbstractEqualityByValTests.cs" />
@@ -108,6 +107,7 @@
     <Compile Include="SLB\StringBinaryWriterTests.cs" />
     <Compile Include="SLB\StreamExtensionsTests.cs" />
     <Compile Include="SLB\UIntExtensionsTests.cs" />
+    <Compile Include="Testing\Matchers.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
